4.21

+++ osa4/controllers/blogs.js
@@ -13,12 +13,10 @@ blogsRouter.post('/', async (request, response) => {
     const body = request.body
 
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    
     if (!decodedToken.id) {
         return response.status(401).json({ error: 'token invalid' })
     }  
     const user = await User.findById(decodedToken.id)
-
 
     const blog = new Blog({
         title: body.title,
@@ -35,8 +33,20 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }  
+    const user = await User.findById(decodedToken.id)
+
+    const blog = await Blog.findById(request.params.id)
+
+    if ( blog.user.toString() === user._id.toString() ){
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    }else{
+        response.status(401).json({ error: 'token invalid' })
+    }
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
